requirements.txt
pymongo==4.6.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.5.3
python-dotenv==1.0.0
